# frontend/openapi/fixture.yaml
openapi: 3.1.0
info:
  title: Amezo API (fixture)
  version: 0.0.0
  description: >
    Placeholder spec so the codegen pipeline (springdoc -> OpenAPI ->
    openapi-typescript) has something to run against before the Spring
    backend exposes its own. Modeled on docs/api-design.md's documented
    GET /products contract, not the backend's current partial
    implementation (q-only as of this writing — see ProductController.java).
    Swap `gen:api` in package.json to point at the live springdoc endpoint
    once the backend catches up to this contract; this file goes away then.
paths:
...
  /api/v1/refund-requests:
    get:
      operationId: listBuyerRefundRequests
      description: The signed-in buyer's own refund requests.
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/RefundStatus' }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundRequestPage'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    post:
      operationId: createRefundRequest
      description: >-
        Buyer session required, and every line must belong to the caller's own
        order. 409 when a line already has an open request, 422 when the return
        window has closed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRefundRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundRequestDetail'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/refund-requests/{refundRequestId}:
    get:
      operationId: getRefundRequest
      description: Readable by the buyer who raised it and by the seller who owes it.
      parameters:
        - name: refundRequestId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundRequestDetail'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    patch:
      operationId: updateRefundRequest
      description: >-
        The state machine lives on the resource rather than on action URLs.

        Seller moves REQUESTED to AWAITING_RETURN (approving a REFUND, which asks
        for the item back), to APPROVED (approving a REPLACEMENT, where there is
        nothing to wait for) or to DECLINED, then on to RETURN_RECEIVED, REFUNDED
        or REPLACEMENT_SENT.

        Seller may also move APPROVED or AWAITING_RETURN back to REQUESTED - the
        "Undo approval" the Seller Refunds design puts beside the release button.
        It clears approvedAmount, approvedAt and returnTrackingNumber, and restores
        requestedResolution, because the approval's own facts must not outlive the
        decision that set them. It is NOT available from RETURN_RECEIVED: once the
        buyer has posted the item back, unwinding the approval would strand it.

        Buyer may only CANCEL, and only while REQUESTED. An illegal transition is
        409, as is a transition the caller's own side of the refund cannot make.

        Note what a REFUNDED transition does and does not do. It records that the
        seller released an amount and when. It moves no money: there is no payment
        table, gateway or ledger in this system, and nothing here pretends
        otherwise. It is also what an order's derived REFUNDED status reads.
      parameters:
        - name: refundRequestId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRefundRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundRequestDetail'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/sellers/me/refund-requests:
    get:
      operationId: sellerListRefundRequests
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/RefundStatus' }
        - name: q
          in: query
          description: Matches buyer name, buyer email, order reference and product title.
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundRequestPage'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/v1/sellers/me/refund-requests/facets:
    get:
      operationId: getSellerRefundFacets
      description: Counts for the refund queue's tabs. Ignores status, as above.
      parameters:
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacetList'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  schemas:
...
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }

    # --- /api/v1 shapes -------------------------------------------------------

    # Supersedes the three-value status on the unversioned seller endpoints.
    # PACKED is the handover step the seller screens add; IN_TRANSIT and
    # OUT_FOR_DELIVERY are what the buyer's tracking timeline shows once a
    # carrier is scanning.
...
    RefundStatus:
      type: string
      enum:
        [REQUESTED, APPROVED, AWAITING_RETURN, RETURN_RECEIVED, REFUNDED, REPLACEMENT_SENT, DECLINED, CANCELLED]
    RefundResolution:
      type: string
      enum: [REFUND, REPLACEMENT]
    # The design offers "the card you paid with" or "that card is closed, contact
    # me" - not store credit, which the marketplace has no concept of.
    RefundPayout:
      type: string
      enum: [ORIGINAL_PAYMENT, ALTERNATE_METHOD]
...
    StoreRef:
      type: object
      required: [name, handle]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        handle: { $ref: '#/components/schemas/StoreHandle' }
...
    RefundRequestLine:
      type: object
      required: [orderLineId, productTitle, variantLabel, quantity, unitPrice, lineTotal]
      properties:
        orderLineId: { type: string, format: uuid }
        productTitle: { type: string }
        variantLabel: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        lineTotal: { type: number }
    RefundRequestSummary:
      type: object
      required: [id, reference, status, resolution, requestedAt, requestedAmount, currency]
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        status: { $ref: '#/components/schemas/RefundStatus' }
        resolution: { $ref: '#/components/schemas/RefundResolution' }
        requestedAt: { type: string, format: date-time }
        requestedAmount: { type: number }
        approvedAmount: { type: number, nullable: true }
        currency: { type: string }
        orderId: { type: string, format: uuid }
        orderReference: { type: string }
        returnTrackingNumber: { type: string, nullable: true }
        # The Seller Refunds table prints a Buyer column and an Items column, so a
        # row carries both rather than being a line of reference codes. Nullable
        # because a guest checkout supplies only an email, and buyer_identity's
        # full_name is nullable.
        buyerName: { type: string, nullable: true }
        buyerEmail: { type: string, nullable: true }
        # "2 x Aurora One Wireless Headphones, 1 x Travel Case" - the Items column,
        # assembled by the server because the product titles live in the catalogue
        # and a summary row carries no lines to build it from.
        items: { type: string, nullable: true }
    RefundRequestDetail:
      type: object
      required:
        [id, reference, status, resolution, payout, detail, requestedAt, requestedAmount, currency, orderId, orderReference, lines]
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        status: { $ref: '#/components/schemas/RefundStatus' }
        resolution: { $ref: '#/components/schemas/RefundResolution' }
        # What the BUYER originally asked for, which `resolution` may no longer be:
        # the seller can settle a replacement request with money or the reverse, and
        # overwriting the ask would lose the fact that they did. The decision
        # panel's "Wants ..." line reads this.
        requestedResolution: { $ref: '#/components/schemas/RefundResolution' }
        payout: { $ref: '#/components/schemas/RefundPayout' }
        detail: { type: string }
        requestedAt: { type: string, format: date-time }
        requestedAmount: { type: number }
        approvedAmount: { type: number, nullable: true }
        currency: { type: string }
        orderId: { type: string, format: uuid }
        orderReference: { type: string }
        orderPlacedAt: { type: string, format: date-time }
        buyerName: { type: string, nullable: true }
        buyerEmail: { type: string, nullable: true }
        seller: { $ref: '#/components/schemas/StoreRef' }
        paymentMethod: { type: string, nullable: true }
        lines:
          type: array
          items: { $ref: '#/components/schemas/RefundRequestLine' }
        # The seller's own account of what happened, newest last.
        events:
          type: array
          items: { $ref: '#/components/schemas/RefundEvent' }
        approvedAt: { type: string, format: date-time, nullable: true }
        declinedAt: { type: string, format: date-time, nullable: true }
        declineReason: { type: string, nullable: true }
        returnReceivedAt: { type: string, format: date-time, nullable: true }
        refundedAt: { type: string, format: date-time, nullable: true }
        replacementSentAt: { type: string, format: date-time, nullable: true }
        returnTrackingNumber: { type: string, nullable: true }
    RefundRequestPage:
      type: object
      required: [content, page, totalElements, totalPages]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/RefundRequestSummary' }
        page: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
    CreateRefundRequest:
      type: object
      required: [orderId, lines, resolution, detail]
      properties:
        orderId: { type: string, format: uuid }
        lines:
          type: array
          minItems: 1
          items:
            type: object
            required: [orderLineId, quantity]
            properties:
              orderLineId: { type: string, format: uuid }
              quantity: { type: integer, minimum: 1 }
        resolution: { $ref: '#/components/schemas/RefundResolution' }
        # Required only when resolution is REFUND.
        payout:
          oneOf:
            - $ref: '#/components/schemas/RefundPayout'
          nullable: true
        detail: { type: string, minLength: 20 }
    UpdateRefundRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/RefundStatus' }
        # The outcome the SELLER settles on, which need not be the one the buyer
        # asked for - a replacement request can be settled with money, and the
        # reverse. Without it the seller's choice could not be recorded at all
        # and had to be inferred from whether an amount was approved.
        resolution: { $ref: '#/components/schemas/RefundResolution' }
        approvedAmount: { type: number }
        declineReason: { type: string }
        returnTrackingNumber: { type: string }
        # Shown to the buyer on the request's timeline. It is recorded against
        # the transition it accompanied - see RefundEvent.
        note: { type: string }
    RefundEvent:
      type: object
      description: >-
        One step in a refund's history. The timeline is a read of these, not a
        reconstruction from scattered timestamps, which is the only way a note
        can be attached to the step it was written for.
      required: [status, at]
      properties:
        status: { $ref: '#/components/schemas/RefundStatus' }
        at: { type: string, format: date-time }
        note: { type: string, nullable: true }
...
    StoreHandle:
      type: string
      description: >-
        The store's URL segment (amezo.com/stores/{handle}). Lowercase letters,
        digits and inner dashes only - never a leading or trailing dash, so the
        printed URL never ends in one. Case-insensitive on lookup and unique
        across stores; PATCH /api/v1/sellers/me/store answers 409 when taken.
      pattern: '^[a-z0-9][a-z0-9-]{0,37}[a-z0-9]$'
      minLength: 2
      maxLength: 39
      example: northwind-supply
...
    Facet:
      type: object
      description: >-
        One tab. `key` is whatever the matching list endpoint accepts for that
        bucket, so a tab needs no translation table to become a filter.
      required: [key, count]
      properties:
        key: { type: string }
        count: { type: integer }
        # Money in this bucket, where the design prints it. Absent where a
        # bucket has no meaningful total rather than sent as zero.
        value: { type: number, nullable: true }
        currency: { type: string, nullable: true }
    FacetList:
      type: object
      required: [facets]
      properties:
        facets:
          type: array
          items: { $ref: '#/components/schemas/Facet' }
...